msm: clock: Keep proper track of current voltage level
Fix a bug where vdd_class->cur_level is not updated when regulator api
calls are successful. The impact of the bug is unnecessary calls to
regulator APIs but not additional RPM calls since the regulator APIs
will optimize out repeated requests for the same voltage/current.

In case the regulator APIs fail, this bug might also result in
undervolting. But that's an unlikely case in a real product.

Change-Id: Ibe3574b6b72fcf898a5bfe9b5b253a8d0e0125ae
Signed-off-by: Patrick Daly <pdaly@codeaurora.org>
Signed-off-by: flar2 <asegaert@gmail.com>